+++ src/components/error_pages/NoAccessPage.tsx
@@ -1,4 +1,4 @@
-import Header from "./Header";
+import Header from "../Header";
 
 export function NoAccessPage() {
     return (
